[utils] Extract createLogOnce to its own module
