+++ packages/utils/src/warn.test.ts
@@ -1,5 +1,5 @@
 import { expect, vi } from 'vitest';
-import { createLogOnce, reset, warn } from './warn';
+import { reset, warn } from './warn';
 import { error } from './error';
 
 describe('warn', () => {
@@ -58,31 +58,5 @@ describe('error', () => {
     error('message');
     expect(warnSpy).toHaveBeenCalledTimes(1);
     expect(errorSpy).toHaveBeenCalledTimes(1);
-  });
-});
-
-describe('createLogOnce', () => {
-  beforeEach(() => {
-    reset();
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it('creates a logger with a custom prefix', () => {
-    const spy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    const logOnce = createLogOnce('warn', 'My Library');
-    logOnce('message');
-    logOnce('message');
-    expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy).toHaveBeenCalledWith('My Library: message');
-  });
-
-  it('creates a logger without a prefix', () => {
-    const spy = vi.spyOn(console, 'error').mockImplementation(() => {});
-    const logOnce = createLogOnce('error');
-    logOnce('message');
-    expect(spy).toHaveBeenCalledWith('message');
   });
 });
